Support Azure OpenAI Service
https://learn.microsoft.com/en-us/azure/ai-services/openai/reference

This has the same basic interface as OpenAI itself, but uses a different
auth header.

I called then new version "AzureOpenai" rather than just "Azure" as
Azure hosts other LLMs apart from OpenAI as well.

This works fine when the .api_configs.toml and prompts.toml are set to
use it. However I did see weirdness when specifying the api on the
commandline.

Even though the new strings are AzureOpenai and azureopenai everywhere,
the commandline needs `--api azure-openai` with a hypen to select it.
I'm guessing that this is from some default camel-case to kebab-case handling
of AzureOpenai.